Removed unaligned view mode triggered by --bit-skip
Previously a bit-skip forced a set number of bits to be dropped from the
start of a viewed track bitstream, with the remainder read without any
further sync processing. This was useful to work around a sync matching bug
that has now been fixed.

The bit skip option is now applied to the regular track view mode, and still
drops bits from the start of the bitstream. The remainder of the stream is
still processed with sync processing enabled. The option allows data from
track start to the first index to be checked for interesting mis-aligned
data.