Added support for Speedlock 5K checksums
Speedlock protections using 8K sectors may have a 16-bit checksum after the
first 5K of data. We now check for this and report it if found. This is
shown as a new 'crc5k2' marker in the -vv output.

Changed the 6K checksums from crc0 to crc6k and crc1 to crc6k2 to better
describe the size of the data they cover.

Changed EDSK reader to only warn about missing checksums if no checksum
method was detected. Previously we'd warn if only 6K of data was stored,
but that's enough for some checksum types.