Add fallback method for DISCiPLE IMG detection
We previously relied on finding the sector chain for the first file sector,
but that fails if the file is too short for a second sector.

We now also check for a non-null header in the directory entry, and whether it
matches the first 9 bytes of the file. That will be present for common CODE and
BASIC file types.